Move draft progress service to its own file
